+++ app/controllers/staff/forecast_uploads_controller.rb
@@ -39,6 +39,11 @@ class Staff::ForecastUploadsController < Staff::BaseController
       @errors = importer.errors
 
       if @errors.empty?
+        imported_forecasts = importer.imported_forecasts.compact
+        @grouped_forecasts = imported_forecasts
+          .map { |forecast| ForecastPresenter.new(forecast) }
+          .group_by { |forecast| ActivityPresenter.new(forecast.parent_activity) }
+        @total_forecast = TotalPresenter.new(imported_forecasts.sum(&:value)).value
         @success = true
         flash.now[:notice] = t("action.forecast.upload.success")
       end